requirement.txt
# PyTorch
torch==2.2.1
torchvision==0.17.1

# Core libs
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.1
seaborn==0.12.2

# Utilities
tqdm==4.66.1
pillow==10.1.0
